Rename pion logger package
