+++ logger/pionlogger/logadapter.go
@@ -1,4 +1,4 @@
-package pion
+package pionlogger
 
 import (
 	"fmt"
